requirements.txt
fastapi
uvicorn
python-multipart
pillow
opencv-python
imutils
reportlab
img2pdf